api/serializers/product_serializers.py
from rest_framework import serializers
from api.models.model_products import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
